Refact: text split animation(캐릭터 대사, 코드 입력) 수정
